refactor: separate booking management logic and UI
- Move booking management logic to useBooking hook
- Create dedicated useBookingManagement hook for business logic
- Implement proper TypeScript types and interfaces
- Fix setTimeout type declaration
- Add filter functionality for booking status
- Improve component structure in booking-management screen
- Add cancel dialog and booking actions
